examples/rag_agent/pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "RAG Agent with behavior-based decomposition for adaptive retrieval strategies"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [{ name = "OpenSymbolicAI" }]
keywords = ["rag", "retrieval", "agent", "llm", "opensymbolicai"]

dependencies = [
    "opensymbolicai-core",
    # Vector store
    "chromadb>=0.4.0",
    # Embeddings
    "sentence-transformers>=2.2.0",
    # Wikipedia data source
    "wikipedia-api>=0.6.0",
    # Environment variables
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
rag-agent = "rag_agent.main:main"
rag-load-wiki = "rag_agent.wikipedia_loader:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rag_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
